+++ Editor/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleTo("Unity.XR.CoreUtils.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.XR.CoreUtils.Editor.Tests")]
 // Shared test assembly used as part of Unity testing conventions.
 [assembly: InternalsVisibleTo("Assembly-CSharp-Editor-testable")]