+++ Runtime/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 #if UNITY_EDITOR
 using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.XR.CoreUtils.Editor")]
-[assembly: InternalsVisibleTo("Unity.XR.CoreUtils.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.XR.CoreUtils.Editor.Tests")]
 // Shared test assembly used as part of Unity testing conventions.
 [assembly: InternalsVisibleTo("Assembly-CSharp-Editor-testable")]
 [assembly: InternalsVisibleTo("Assembly-CSharp-testable")]