+++ Runtime/Collections/ReadOnlyListSpan.cs
@@ -1,19 +1,22 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Text;
 
 namespace Unity.XR.CoreUtils.Collections
 {
     /// <summary>
-    /// Wraps a <see cref="List{T}"/> to provide a read-only view of its memory without copying any elements.
-    /// It is preferable to use this collection in API designs instead of `IReadOnlyCollection` because
-    /// <see cref="GetEnumerator"/> returns a value-type enumerator and does not perform any heap allocations. Additionally, this collection has the ability to provide a slice of a list.
+    /// Wraps a list or array to provide a read-only view of some or all elements. Elements are not copied, so if the
+    /// underlying collection changes, the `ReadOnlyListSpan` will see the updated elements.
     /// </summary>
     /// <remarks>
+    /// It is preferable to use this collection in API designs instead of `IReadOnlyCollection` because
+    /// <see cref="GetEnumerator"/> returns a value-type enumerator and does not perform any heap allocations.
+    ///
     /// This collection is not thread-safe.
     /// </remarks>
     /// <typeparam name="T">The element type.</typeparam>
-    public struct ReadOnlyListSpan<T> : IReadOnlyList<T>
+    public struct ReadOnlyListSpan<T> : IReadOnlyList<T>, IEquatable<ReadOnlyListSpan<T>>
     {
         static ReadOnlyListSpan<T> s_EmptyList = new();
         Enumerator m_Enumerator;
@@ -45,7 +48,7 @@ namespace Unity.XR.CoreUtils.Collections
         /// </summary>
         /// <param name="list">The list to wrap.</param>
         /// <exception cref="ArgumentNullException">Thrown if <paramref name="list"/> is <see langword="null"/>.</exception>
-        public ReadOnlyListSpan(List<T> list)
+        public ReadOnlyListSpan(IReadOnlyList<T> list)
         {
             if (list == null)
                 throw new ArgumentNullException(nameof(list));
@@ -62,7 +65,7 @@ namespace Unity.XR.CoreUtils.Collections
         /// <exception cref="ArgumentNullException">Thrown if <paramref name="list"/> is <see langword="null"/>.</exception>
         /// <exception cref="ArgumentOutOfRangeException"> Thrown if
         /// start or length are outside the bounds of the list.</exception>
-        public ReadOnlyListSpan(List<T> list, int start, int length)
+        public ReadOnlyListSpan(IReadOnlyList<T> list, int start, int length)
         {
             if (list == null)
                 throw new ArgumentNullException(nameof(list));
@@ -94,19 +97,13 @@ namespace Unity.XR.CoreUtils.Collections
         /// Returns an empty read-only list with the specified type argument.
         /// </summary>
         /// <returns>The empty read-only list.</returns>
-        public static ReadOnlyListSpan<T> Empty()
-        {
-            return s_EmptyList;
-        }
+        public static ReadOnlyListSpan<T> Empty() => s_EmptyList;
 
         /// <summary>
         /// Returns an enumerator that iterates through the read-only list.
         /// </summary>
         /// <returns>The enumerator.</returns>
-        public Enumerator GetEnumerator()
-        {
-            return m_Enumerator;
-        }
+        public Enumerator GetEnumerator() => m_Enumerator;
 
         /// <summary>
         /// Returns an enumerator that iterates through the read-only list.
@@ -117,10 +114,7 @@ namespace Unity.XR.CoreUtils.Collections
         /// > Use the public <see cref="GetEnumerator"/> overload instead.
         /// </remarks>
         /// <returns>The boxed enumerator.</returns>
-        IEnumerator<T> IEnumerable<T>.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
+        IEnumerator<T> IEnumerable<T>.GetEnumerator() => GetEnumerator();
 
         /// <summary>
         /// Returns an enumerator that iterates through the read-only list.
@@ -131,9 +125,71 @@ namespace Unity.XR.CoreUtils.Collections
         /// > Use the public <see cref="GetEnumerator"/> overload instead.
         /// </remarks>
         /// <returns>The boxed enumerator.</returns>
-        IEnumerator IEnumerable.GetEnumerator()
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+        /// <summary>
+        /// Indicates whether the current object is equal to another object of the same type.
+        /// </summary>
+        /// <param name="other">An object to compare with this object.</param>
+        /// <returns>`true` if the current object is equal to the `other` parameter. Otherwise, `false`.</returns>
+        /// <remarks>
+        /// Two instances compare equal if they are read-only views of the same list, for the same start and end indices.
+        /// </remarks>
+        public bool Equals(ReadOnlyListSpan<T> other)
         {
-            return GetEnumerator();
+            return ReferenceEquals(m_Enumerator.list, other.m_Enumerator.list)
+                && m_Enumerator.start == other.m_Enumerator.start
+                && m_Enumerator.end == other.m_Enumerator.end;
+        }
+
+        /// <summary>
+        /// Indicates whether the current object is equal to another object of the same type.
+        /// </summary>
+        /// <param name="obj">An object to compare with this object.</param>
+        /// <returns>`true` if the current object is equal to the `other` parameter. Otherwise, `false`.</returns>
+        public override bool Equals(object obj)
+            => obj is ReadOnlyListSpan<T> other && Equals(other);
+
+        /// <summary>
+        /// Returns `true` if objects are equal by <see cref="Equals(Unity.XR.CoreUtils.Collections.ReadOnlyListSpan{T})"/>.
+        /// Otherwise, `false`.
+        /// </summary>
+        /// <param name="lhs">The left-hand side of the comparison.</param>
+        /// <param name="rhs">The right-hand side of the comparison.</param>
+        /// <returns>`true` if objects are equal. Otherwise, `false`.</returns>
+        public static bool operator ==(ReadOnlyListSpan<T> lhs, ReadOnlyListSpan<T> rhs) => lhs.Equals(rhs);
+
+        /// <summary>
+        /// Returns `false` if objects are equal by <see cref="Equals(Unity.XR.CoreUtils.Collections.ReadOnlyListSpan{T})"/>.
+        /// Otherwise, `true`.
+        /// </summary>
+        /// <param name="lhs">The left-hand side of the comparison.</param>
+        /// <param name="rhs">The right-hand side of the comparison.</param>
+        /// <returns>`false` if objects are equal. Otherwise, `true`.</returns>
+        public static bool operator !=(ReadOnlyListSpan<T> lhs, ReadOnlyListSpan<T> rhs) => !(lhs == rhs);
+
+        /// <summary>
+        /// Get a hash code for this object.
+        /// </summary>
+        /// <returns>The hash code.</returns>
+        public override int GetHashCode()
+            => HashCode.Combine(m_Enumerator.list, m_Enumerator.start, m_Enumerator.end);
+
+        /// <summary>
+        /// Returns a string that represents the current object.
+        /// </summary>
+        /// <returns>The string.</returns>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.AppendLine("{");
+            for (var i = m_Enumerator.start; i < m_Enumerator.end; i++)
+            {
+                var item = m_Enumerator.list[i];
+                sb.AppendLine(item == null ? "  null," : $"  {m_Enumerator.list[i].ToString()},");
+            }
+            sb.Append("}");
+            return sb.ToString();
         }
 
         /// <summary>
@@ -168,23 +224,19 @@ namespace Unity.XR.CoreUtils.Collections
             }
 
             object IEnumerator.Current => Current;
-            internal List<T> list;
+            internal IReadOnlyList<T> list;
             int m_CurrentIndex;
 
-            /// <summary>
-            /// Provides an enumerator for the elements of a `ReadOnlyListSpan`.
-            /// </summary>
-            /// <param name="list"></param>
-            internal Enumerator(List<T> list) : this(list, 0, list.Count) { }
+            internal Enumerator(IReadOnlyList<T> list) : this(list, 0, list.Count) { }
 
             /// <summary>
-            /// Provides an enumerator for a slice of the elements of a `ReadOnlyListSpan` beginning with the
+            /// Provides an enumerator for a slice of the elements of a list beginning with the
             /// `start` index and ending at the `end` index.
             /// </summary>
             /// <param name="list">The list to enumerate.</param>
             /// <param name="start">The zero-based index at which to begin this slice.</param>
             /// <param name="end">The desired zero-based index at which to end the slice.</param>
-            internal Enumerator(List<T> list, int start, int end)
+            internal Enumerator(IReadOnlyList<T> list, int start, int end)
             {
                 this.list = list;
                 this.start = start;
@@ -205,10 +257,7 @@ namespace Unity.XR.CoreUtils.Collections
             /// <summary>
             /// Sets the enumerator to its initial position, which is before the first element in the collection.
             /// </summary>
-            public void Reset()
-            {
-                m_CurrentIndex = start - 1;
-            }
+            public void Reset() => m_CurrentIndex = start - 1;
 
             void IDisposable.Dispose() { }
         }