+++ Tests/Editor/BoundsUtilsTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEngine;
 
-namespace Unity.XR.CoreUtils.EditorTests
+namespace Unity.XR.CoreUtils.Editor.Tests
 {
     [TestFixture]
     class BoundsUtilsTests