+++ Tests/Editor/EditorPrefsUtilsTests.cs
@@ -7,7 +7,7 @@ using Unity.XR.CoreUtils.Editor;
 using UnityEditor;
 using UnityEngine;
 
-namespace Unity.XR.CoreUtils.EditorTests
+namespace Unity.XR.CoreUtils.Editor.Tests
 {
     class EditorPrefsUtilsTests
     {